# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "commerce-system API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/members": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "모든 회원의 목록을 조회합니다. (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "회원 목록 조회",
                "responses": {
                    "200": {
                        "description": "회원 목록",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.MemberResponse"
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "목록 조회 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "새로운 회원을 등록합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "회원 가입",
                "parameters": [
                    {
                        "description": "회원 가입 정보",
                        "name": "registerRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "가입 성공",
                        "schema": {
                            "$ref": "#/definitions/response.MemberResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/members/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "인증된 사용자의 정보를 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "내 정보 조회",
                "responses": {
                    "200": {
                        "description": "내 정보",
                        "schema": {
                            "$ref": "#/definitions/response.MemberResponse"
                        }
                    },
                    "500": {
                        "description": "정보 조회 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "인증된 사용자의 정보를 수정합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "내 정보 수정",
                "parameters": [
                    {
                        "description": "수정할 정보",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "수정 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "수정 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "인증된 사용자의 계정을 삭제합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "회원 탈퇴",
                "responses": {
                    "200": {
                        "description": "삭제 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "삭제 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/members/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "특정 월의 회원 가입 통계를 조회합니다. (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "members"
                ],
                "summary": "회원 통계 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "조회할 월 (YYYY-MM)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "통계 정보",
                        "schema": {
                            "$ref": "#/definitions/response.MemberStatsResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "통계 조회 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.RegisterMemberRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "member_number": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "request.UpdateMemberRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "response.MemberResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "is_withdrawn": {
                    "type": "boolean"
                },
                "member_number": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "withdrawn_at": {
                    "type": "string"
                }
            }
        },
        "response.MemberStatsResponse": {
            "type": "object",
            "properties": {
                "deleted_members": {
                    "type": "integer"
                },
                "joined_members": {
                    "type": "integer"
                },
                "month": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "A commerce-system service API in Go using Gin framework",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}